backend/pyproject.toml
[project]
name = "nouvaris-agents-v2"
version = "0.1.0"
description = "Nouvaris Agents V2"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "pydantic>=2",
    "python-dotenv",
    "pyyaml",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "redis[hiredis]>=5.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["app/tests", "tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"



